+++ spacekit/skopes/hst/svm/corrupt.py
@@ -16,9 +16,16 @@ from astropy.io import fits
 import time
 from tqdm import tqdm
 from progressbar import ProgressBar
-from drizzlepac import runsinglehap
+
 from spacekit.generator.draw import DrawMosaics
 from spacekit.analyzer.track import stopwatch
+
+try:
+    from drizzlepac import runsinglehap
+except ImportError:
+    print(
+        "Running this script requires drizzlepac to be installed. Run `pip install drizzlepac` then try again."
+    )
 
 SVM_QUALITY_TESTING = "on"
 os.environ["SVM_QUALITY_TESTING"] = "on"